feat: move RelayListCollection and RelaySection to commons
Move the RelayListCollection and RelaySection data classes to commons.
The sections() extension function stays in Android as it depends on
R.string resource IDs.
